# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для системы осмотра жилых помещений ЖКХ.\nСистема предназначена для координации работы инспекторов,\nуправления заданиями на осмотр зданий и формирования отчётности.",
        "title": "JKH Inspection API",
        "contact": {
            "name": "Техническая поддержка",
            "email": "support@jkh.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/elements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех элементов для чек-листов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочник элементов"
                ],
                "summary": "Получить список элементов справочника",
                "responses": {
                    "200": {
                        "description": "Список элементов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ElementCatalogResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создание нового элемента для использования в чек-листах (например: Фундамент, Кровля)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочник элементов"
                ],
                "summary": "Создать элемент справочника",
                "parameters": [
                    {
                        "description": "Данные элемента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateElementCatalogRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Элемент успешно создан",
                        "schema": {
                            "$ref": "#/definitions/models.ElementCatalogResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Элемент с таким именем уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/elements/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает информацию о конкретном элементе справочника",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочник элементов"
                ],
                "summary": "Получить элемент по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные элемента",
                        "schema": {
                            "$ref": "#/definitions/models.ElementCatalogResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Элемент не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление данных элемента справочника",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочник элементов"
                ],
                "summary": "Обновить элемент",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateElementCatalogRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленные данные элемента",
                        "schema": {
                            "$ref": "#/definitions/models.ElementCatalogResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Элемент не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Имя элемента уже занято",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаление элемента из справочника",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Справочник элементов"
                ],
                "summary": "Удалить элемент",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID элемента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Элемент успешно удален"
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Элемент не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Элемент используется в чек-листах",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateElementCatalogRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "category": {
                    "description": "Категория элемента (опционально, для фильтрации и группировки).\nУказатель (*string) позволяет отличить \"не передано\" от \"пустая строка\".\nomitempty — если поле nil, оно не включается в JSON-ответ.",
                    "type": "string"
                },
                "name": {
                    "description": "Название элемента (например, \"Фундамент\", \"Кровля\", \"Стены\").\nПоле обязательно (binding:\"required\") — валидация на уровне Gin.",
                    "type": "string"
                }
            }
        },
...
        "models.ElementCatalogResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Категория (всегда строка, даже если пустая)",
                    "type": "string"
                },
                "id": {
                    "description": "Уникальный идентификатор элемента",
                    "type": "integer"
                },
                "name": {
                    "description": "Название элемента",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите JWT токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}